+++ src/test/groovy/no/skatteetaten/aurora/boober/model/AuroraDeploymentSpecBuilderTest.groovy
@@ -1,0 +1,26 @@
+package no.skatteetaten.aurora.boober.model
+
+import no.skatteetaten.aurora.boober.mapper.v1.AuroraDeploymentSpecBuilderKt
+import no.skatteetaten.aurora.boober.service.AuroraConfigException
+
+class AuroraDeploymentSpecBuilderTest extends AbstractAuroraConfigTest {
+
+  static AuroraDeploymentSpec createDeploymentSpec(Map<String, String> auroraConfigJson, ApplicationId applicationId) {
+
+    AuroraConfig auroraConfig = createAuroraConfig(auroraConfigJson)
+    AuroraDeploymentSpecBuilderKt.createAuroraDeploymentSpec(auroraConfig, applicationId, "", [], [:])
+  }
+
+  def auroraConfigJson = defaultAuroraConfig()
+
+  def "Fails when envFile does not start with about"() {
+    given:
+      auroraConfigJson["utv/aos-simple.json"] = '''{ "envFiles": "foo" }'''
+
+    when:
+      createDeploymentSpec(auroraConfigJson, DEFAULT_AID)
+
+    then:
+      thrown(AuroraConfigException)
+  }
+}
