+++ src/test/groovy/no/skatteetaten/aurora/boober/service/AuroraConfigFieldTest.groovy
@@ -14,6 +14,7 @@ import no.skatteetaten.aurora.boober.controller.security.User
 import no.skatteetaten.aurora.boober.controller.security.UserDetailsProvider
 import no.skatteetaten.aurora.boober.mapper.v1.AuroraVolumeMapperV1
 import no.skatteetaten.aurora.boober.model.ApplicationId
+import no.skatteetaten.aurora.boober.model.AuroraConfigHelperKt
 import no.skatteetaten.aurora.boober.service.openshift.OpenShiftClient
 import spock.lang.Specification
 import spock.mock.DetachedMockFactory
