+++ src/test/groovy/no/skatteetaten/aurora/boober/service/DeployBundleServiceTest.groovy
@@ -8,6 +8,7 @@ import com.fasterxml.jackson.databind.node.ObjectNode
 import no.skatteetaten.aurora.boober.model.ApplicationId
 import no.skatteetaten.aurora.boober.model.AuroraConfig
 import no.skatteetaten.aurora.boober.model.AuroraConfigFile
+import no.skatteetaten.aurora.boober.model.AuroraConfigHelperKt
 
 class DeployBundleServiceTest extends AbstractMockedOpenShiftSpecification {
