+++ src/test/kotlin/no/skatteetaten/aurora/boober/model/AuroraConfigHelper.kt
@@ -1,10 +1,8 @@
-package no.skatteetaten.aurora.boober.service
+package no.skatteetaten.aurora.boober.model
+
 
 import com.fasterxml.jackson.databind.JsonNode
 import no.skatteetaten.aurora.boober.Configuration
-import no.skatteetaten.aurora.boober.model.ApplicationId
-import no.skatteetaten.aurora.boober.model.AuroraConfig
-import no.skatteetaten.aurora.boober.model.AuroraConfigFile
 import no.skatteetaten.aurora.boober.utils.findAllPointers
 import java.io.File
 
@@ -14,9 +12,9 @@ val folder = File(AuroraConfigHelper::class.java.getResource("/samples/config").
 
 fun getAuroraConfigSamples(): AuroraConfig {
     val files = folder.walkBottomUp()
-        .onEnter { it.name != "secret" }
-        .filter { it.isFile }
-        .associate { it.relativeTo(folder).path to it }
+            .onEnter { it.name != "secret" }
+            .filter { it.isFile }
+            .associate { it.relativeTo(folder).path to it }
 
     val nodes = files.map {
         it.key to convertFileToJsonNode(it.value)
@@ -38,24 +36,24 @@ fun createAuroraConfig(aid: ApplicationId, affiliation: String = "aos", addition
 fun getSampleFiles(aid: ApplicationId, additionalFile: String? = null): Map<String, JsonNode?> {
 
     return collectFilesToMapOfJsonNode(
-        "about.json",
-        "${aid.application}.json",
-        "${aid.environment}/about.json",
-        "${aid.environment}/${aid.application}.json",
-        additionalFile?.let { it } ?: ""
+            "about.json",
+            "${aid.application}.json",
+            "${aid.environment}/about.json",
+            "${aid.environment}/${aid.application}.json",
+            additionalFile?.let { it } ?: ""
     )
 }
 
 fun getResultFiles(aid: ApplicationId): Map<String, JsonNode?> {
     val baseFolder = File(AuroraConfigHelper::class.java
-        .getResource("/samples/result/${aid.environment}/${aid.application}").file)
+            .getResource("/samples/result/${aid.environment}/${aid.application}").file)
 
     return getFiles(baseFolder)
 }
 
 fun getDeployResultFiles(aid: ApplicationId): Map<String, JsonNode?> {
     val baseFolder = File(AuroraConfigHelper::class.java
-        .getResource("/samples/deployresult/${aid.environment}/${aid.application}").file)
+            .getResource("/samples/deployresult/${aid.environment}/${aid.application}").file)
 
     return getFiles(baseFolder, aid.application)
 }
@@ -84,5 +82,4 @@ private fun convertFileToJsonNode(file: File): JsonNode? {
     val mapper = Configuration().mapper()
     return mapper.readValue(file, JsonNode::class.java)
 }
-
